rename protocolMachine to machine
